+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
 * Update icons.
 * Update QML style to match with OS Theme.
+* Get Matrix Server address from user input in Login and CIBA Login pages.
 
 
 ## [0.0.7]